@@ -5,11 +5,11 @@
   "dependencies": {
     "@unicef/react-org-chart": "file:..",
     "gh-pages": "^2.2.0",
-    "react": "^16.12.0",
+    "react": "^18.3.1",
     "react-d3": "^0.4.0",
-    "react-dom": "^16.12.0",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.3.0"
+    "react-scripts": "^5.0.1"
   },
   "homepage": "https://unicef.github.io/react-org-chart",
   "repository": {
