@@ -1,16 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
+import { createRoot } from 'react-dom/client';
 
 test('renders App without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(App, div);
-  ReactDOM.unmountComponentAtNode(div);
+  const root = createRoot(document.createElement('root'));
+  root.render(<App />);
 });
 
 test('contains a react-org-chart div', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
-  expect(div.innerHTML).toContain('<div id=\"react-org-chart\"></div>');
+  const root = createRoot(document.createElement('root'));
+  root.render(<App />);
+  expect(root).toContain('<div id=\"react-org-chart\" data-testid=\"react-org-chart\"></div>');
   ReactDOM.unmountComponentAtNode(div);
 });
